+++ src/Exceptions/WatiException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Wati\Http\Exceptions;
+
+use Exception;
+
+class WatiException extends Exception {}